requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
apscheduler==3.6.3
    # via python-telegram-bot
arabic-reshaper==2.1.4
    # via xhtml2pdf
asgiref==3.5.2
    # via django
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   pyhanko
    #   pyhanko-certvalidator
async-timeout==4.0.2
    # via redis
autopep8==1.7.0
    # via -r requirements.in
backports-zoneinfo==0.2.1
    # via
    #   django
    #   django-celery-beat
    #   django-timezone-field
    #   pytz-deprecation-shim
    #   tzlocal
beautifulsoup4==4.11.1
    # via readtime
billiard==3.6.4.0
    # via celery
boto3==1.24.89
    # via -r requirements.in
botocore==1.27.89
    # via
    #   boto3
    #   s3transfer
cachetools==4.2.2
    # via python-telegram-bot
celery==5.2.7
    # via
    #   django-celery-beat
    #   django-celery-results
certifi==2022.9.24
    # via
    #   python-telegram-bot
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pyhanko
    #   yake
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coverage==6.5.0
    # via -r requirements.in
crispy-bootstrap5==0.6
    # via -r requirements.in
cryptography==38.0.1
    # via
    #   pyhanko
    #   pyhanko-certvalidator
cssselect==1.1.0
    # via pyquery
cssselect2==0.7.0
    # via svglib
deepl==1.11.0
    # via -r requirements.in
deprecated==1.2.13
    # via redis
django==4.0.8
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-celery-beat
    #   django-cors-headers
    #   django-htmx
    #   django-minify-html
    #   django-newsfeed
    #   django-ranged-response
    #   django-simple-captcha
    #   django-slugger
    #   django-storages
    #   django-taggit
django-analytical==3.1.0
    # via -r requirements.in
django-celery-beat==2.3.0
    # via -r requirements.in
django-celery-results==2.4.0
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-htmx==1.9.0
    # via -r requirements.in
django-minify-html==1.3.0
    # via -r requirements.in
django-nested-inline==0.4.5
    # via -r requirements.in
django-newsfeed==0.8.8
    # via -r requirements.in
django-ranged-response==0.2.0
    # via django-simple-captcha
django-simple-captcha==0.5.17
    # via -r requirements.in
django-slugger==1.1.2
    # via -r requirements.in
django-social-share==2.2.1
    # via -r requirements.in
django-storages==1.13.1
    # via -r requirements.in
django-taggit==2.1.0
    # via -r requirements.in
django-timezone-field==5.0
    # via django-celery-beat
flake8==5.0.4
    # via -r requirements.in
future==0.18.2
    # via arabic-reshaper
gunicorn==20.1.0
    # via -r requirements.in
hiredis==2.0.0
    # via -r requirements.in
html5lib==1.1
    # via xhtml2pdf
idna==3.4
    # via requests
importlib-metadata==5.0.0
    # via markdown
jellyfish==0.9.0
    # via yake
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.2.4
    # via celery
lxml==4.9.1
    # via
    #   pyquery
    #   svglib
markdown==3.4.1
    # via -r requirements.in
markdown2==2.4.5
    # via readtime
mccabe==0.7.0
    # via flake8
minify-html==0.10.2
    # via django-minify-html
networkx==2.8.7
    # via yake
numpy==1.23.3
    # via yake
oauthlib==3.2.1
    # via
    #   requests-oauthlib
    #   tweepy
oscrypto==1.3.0
    # via pyhanko-certvalidator
packaging==21.3
    # via redis
pillow==9.1.1
    # via
    #   django-simple-captcha
    #   reportlab
    #   xhtml2pdf
prompt-toolkit==3.0.29
    # via click-repl
psycopg2-binary==2.9.4
    # via -r requirements.in
pycodestyle==2.9.1
    # via
    #   autopep8
    #   flake8
pycparser==2.21
    # via cffi
pyflakes==2.5.0
    # via flake8
pyhanko==0.14.0
    # via xhtml2pdf
pyhanko-certvalidator==0.19.5
    # via
    #   pyhanko
    #   xhtml2pdf
pyparsing==3.0.9
    # via packaging
pypdf3==1.0.6
    # via xhtml2pdf
pyquery==1.4.3
    # via readtime
python-bidi==0.4.2
    # via xhtml2pdf
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   botocore
    #   python-crontab
python-dotenv==0.21.0
    # via -r requirements.in
python-telegram-bot==13.14
    # via -r requirements.in
pytz==2022.1
    # via
    #   apscheduler
    #   celery
    #   django-timezone-field
    #   pyhanko
    #   python-telegram-bot
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via pyhanko
qrcode==7.3.1
    # via pyhanko
readtime==2.0.0
    # via -r requirements.in
redis==4.3.4
    # via -r requirements.in
regex==2022.9.13
    # via segtok
reportlab==3.6.11
    # via
    #   svglib
    #   xhtml2pdf
requests==2.28.1
    # via
    #   deepl
    #   pyhanko
    #   pyhanko-certvalidator
    #   requests-oauthlib
    #   tweepy
requests-oauthlib==1.3.1
    # via tweepy
s3transfer==0.6.0
    # via boto3
segtok==1.5.11
    # via yake
six==1.16.0
    # via
    #   apscheduler
    #   click-repl
    #   html5lib
    #   python-bidi
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
svglib==1.4.1
    # via xhtml2pdf
tabulate==0.9.0
    # via yake
tinycss2==1.1.1
    # via
    #   cssselect2
    #   svglib
toml==0.10.2
    # via autopep8
tornado==6.1
    # via python-telegram-bot
tqdm==4.64.1
    # via pypdf3
tweepy==4.10.1
    # via -r requirements.in
tzdata==2022.1
    # via
    #   django-celery-beat
    #   pytz-deprecation-shim
tzlocal==4.2
    # via
    #   apscheduler
    #   pyhanko
unidecode==1.3.6
    # via django-slugger
uritools==4.0.0
    # via pyhanko-certvalidator
urllib3==1.26.12
    # via
    #   botocore
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wrapt==1.14.1
    # via deprecated
xhtml2pdf==0.2.8
    # via -r requirements.in
yake==0.4.8
    # via -r requirements.in
zipp==3.9.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
